Add bounce response system
Enemies need a way to respond to being bounced on, such as by dying, awarding points, etc.
BounceRegistrar serves as a response manager to BounceDetector bounce events.
Other components, like Bounce Rewarder, that want to respond to bounce events can register their events with a BounceRegistrar.

The BounceDetector means of emitting a bounce event was adjusted to provide more metadata about the event.