Remove conflicting uses of readonly
Applying the readonly keyword to a field prevents serializing changes
in the inspector, but the field still appears editable (only those edits
will not be saved). If the field should be editable in the inspector, it
cannot be marked readonly.

It appears that ReactiveProperties do not have the same problem,
but do throw an error in Odin serialization which I have not fully
investigated, so for now those properties for which it may still be
desirable to allow inspector editing have also been unflagged.