Add player death, ending a run, & persistent data
Player may pass through the ExitGate to voluntarily end a run, and in
so doing will keep the points they have earned this run. If they die
before doing this, they do not gain any points from this run (but
they do retain unspent points from previous runs).

The Shrinker was enlarged to prevent the player from being able to
bounce on it from a regular jump from the ground.