Add MovementLimiter and stamina system
NOTE: The stamina usage based on only grounded jumping does not work as
intended because the CharacterController's IsGrounded variable appears
to be reset every frame, causing inaccurate event subscriptions. As
this is the second feature that is broken by the CharacterController's
unconventional and hidden implementation (its skin-based collider makes
contacting other objects more complicated), it should be discarded in
favor of a more reliable movement system.